+++ src/frontend/app/(pages)/film/page.tsx
@@ -1,4 +1,4 @@
-import { fetchFilmList } from '@/lib/dataFetching';
+import { fetchFilmList } from '@/lib/api/dataFetching';
 import { paginate } from '@/lib/helpers/pagination';
 import { Pagination } from '@/components/custom/pagination';
 import { PageTitle } from '@/components/custom/page-title';
