+++ src/frontend/app/(pages)/forecast/page.tsx
@@ -4,7 +4,7 @@ import { PageContent } from '@/components/custom/page-content';
 import { Topline } from '@/interfaces/BoxOffice';
 import { Metadata } from 'next';
 import addDays from 'date-fns/addDays';
-import { fetchForecast } from '@/lib/dataFetching';
+import { fetchForecast } from '@/lib/api/dataFetching';
 
 export const metadata: Metadata = {
 	title: 'Forecast | Box Office Data',
