+++ src/frontend/app/(pages)/search/page.tsx
@@ -5,7 +5,7 @@ import { Country } from '@/interfaces/Country';
 
 import { paginate } from '@/lib/helpers/pagination';
 import { toTitleCase } from '@/lib/helpers/toTitleCase';
-import { fetchSearch } from '@/lib/dataFetching';
+import { fetchSearch } from '@/lib/api/dataFetching';
 
 import { FilmsTable } from '@/components/tables/films-table';
 import { Searchbar } from '@/components/search';
