+++ src/frontend/app/(pages)/time/[year]/q/[quarter]/page.tsx
@@ -2,7 +2,7 @@ import { TimePage } from '@/app/(pages)/time/time';
 import {
 	fetchBoxOfficeInfinite,
 	fetchBoxOfficeSummary,
-} from '@/lib/dataFetching';
+} from '@/lib/api/dataFetching';
 import { getLastDayofMonth } from '@/lib/helpers/dates';
 
 export async function generateMetadata({
