+++ src/frontend/components/tables/distributor-films-table.tsx
@@ -1,4 +1,4 @@
-import { fetchDistributorFilms } from '@/lib/dataFetching';
+import { fetchDistributorFilms } from '@/lib/api/dataFetching';
 import { FilmTable } from '@/components/tables/film-table';
 import { Pagination } from '@/components/custom/pagination';
 import { paginate } from '@/lib/helpers/pagination';
